rename: 폴더 경로 수정
- 파일 추가로 gitkeep 제거
