+++ src/main/java/com/nexters/sseotdabwa/api/users/controller/UserControllerSpec.java
@@ -1,7 +1,8 @@
-package com.nexters.sseotdabwa.api.users;
+package com.nexters.sseotdabwa.api.users.controller;
 
 import io.swagger.v3.oas.annotations.tags.Tag;
 
 @Tag(name = "Users", description = "사용자 API")
 public interface UserControllerSpec {
+
 }
